native: make requests interrupt ongoing dispatch
Sending requests on a different fiber from the fiber running the
dispatch loop (e.g. from a REPL) currently does not result in the
requests being written to the Wayland fd until the server sends some
event that wakes the dispatching fiber.

This commit makes sending requests on one fiber while another is
dispatching interrupt the dispatch, resulting in requests getting sent
promptly without an unnecessary wait on the server or similar.