native: rename listener to handler, simplify
Passing the proxy and user data to the event handler is very C-brained.
High level languages like Janet have closures.

Separating the user data thing from the event handler completely makes
the API simpler and easier to understand. Also, I can in theory get rid
of the user data API without breaking the event handlers in the future
if I find a more satifying way for the user to obtain their data
corresponding to an object in an event arg.